+++ plugins/daytona/src/components/WorkspaceComponent/WorkspaceComponent.tsx
@@ -1,19 +1,37 @@
 import React from 'react';
 import { Grid } from '@material-ui/core';
 import {
-  Header,
-  Page,
   Content,
-  HeaderLabel,
+  TabbedLayout,
 } from '@backstage/core-components';
 import { WorkspaceListCard } from '../WorkspaceListComponent';
+import { AdminCard } from '../AdminCard';
+import { useAccessToken } from '../../hooks';
 
-export const WorkspaceComponent = () => (
-  <Page themeId="tool">
-    <Header title="Welcome to Daytona!">
-      <HeaderLabel label="Owner" value="Daytona" />
-      <HeaderLabel label="Lifecycle" value="Alpha" />
-    </Header>
+export const WorkspaceComponent = () => {
+  const userApiRoles = useAccessToken().apiRoles;
+  const isAdmin = userApiRoles.includes('owner') || userApiRoles.includes('admin') || userApiRoles.includes('user-admin');
+
+  const adminContent = (
+    <TabbedLayout>
+      <TabbedLayout.Route path='workspaces' title="Workspaces">
+          <Grid container spacing={3} direction="column">
+            <Grid item>
+              <WorkspaceListCard />
+            </Grid>
+          </Grid>
+      </TabbedLayout.Route>
+      <TabbedLayout.Route path='admin' title="Admin">
+          <Grid container spacing={3} direction="column">
+            <Grid item>
+              <AdminCard />
+            </Grid>
+          </Grid>
+      </TabbedLayout.Route>
+    </TabbedLayout>
+  );
+
+  const nonAdminContent = (
     <Content>
       <Grid container spacing={3} direction="column">
         <Grid item>
@@ -21,5 +39,11 @@ export const WorkspaceComponent = () => (
         </Grid>
       </Grid>
     </Content>
-  </Page>
-);
+  );
+
+  return (
+    <>
+      { isAdmin ? adminContent : nonAdminContent }
+    </>
+  )
+}